Scale the chart
